+++ langgraph/src/constants.ts
@@ -7,6 +7,8 @@ export const TAG_HIDDEN = "langsmith:hidden";
 
 export const TASKS = "__pregel_tasks";
 
+export const CHECKPOINT_NAMESPACE_SEPARATOR = "|";
+
 export interface SendInterface {
   node: string;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
